+++ front/src/components/Grid/Grid.css
@@ -1,0 +1,30 @@
+h1 {
+    color: orange;
+}
+
+table.grid, .empty {
+    border: solid 1px mediumaquamarine;
+}
+
+table.grid {
+    margin-left: 10px;
+    margin-top: 10px;
+    padding-top: 50px;
+    border-collapse: collapse;
+    border-right: none;
+    border-bottom: none;
+}
+
+tr, td {
+    margin: 0;
+    padding: 0;
+}
+
+.empty {
+    height: 60px;
+    width: 60px;
+    border-top: none;
+    border-left: none;
+    background-color: whitesmoke;
+}
+
