+++ front/src/components/Grid/Grid.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './css/Components.css';
+import './Grid.css'
 
 export default class Grid extends React.Component {
   constructor(props) {
@@ -12,21 +12,21 @@ export default class Grid extends React.Component {
 
   render() {
     const table = [];
-    for (let i = 0; i < this.state.height; i += 1) {
+    for (let i = 0; i < this.state.height; i++) {
       const children = [];
-      for (let j = 0; j < this.state.width; j += 1) {
+      for (let j = 0; j < this.state.width; j++) {
         children.push(
           <td key={j}>
             <div className="empty" />
           </td>,
         );
       }
-
       table.push(<tr key={i}>{children}</tr>);
     }
 
     return (
       <div>
+        <h1>Revolución Industrial</h1>
         <table className="grid">
           <tbody>
             {table}
