+++ front/src/components/ToolBox/Actions/Action.jsx
@@ -1,4 +1,3 @@
-import '../css/Components.css';
 import ToolboxComponent from '../ToolboxComponent';
 
 export default class Action extends ToolboxComponent {
