+++ front/src/components/ToolBox/Actions/remove.png
