+++ front/src/components/ToolBox/Actions/rotate.png
