+++ front/src/components/ToolBox/Machines/Transporter.jsx
@@ -1,5 +1,4 @@
-import transporter from '../../gallery/transporter.png';
-import '../css/Components.css';
+import transporter from './transporter.png';
 import Machine from './Machine';
 
 export default class Transporter extends Machine {
