+++ front/src/components/ToolBox/Machines/crafter.png
