+++ front/src/components/ToolBox/Machines/furnace.png
