+++ front/src/components/ToolBox/Machines/starter.png
