--- front/src/components/css/Components.css
@@ -1,60 +1,0 @@
-.action {
-  width: 60px;
-  height: 60px;
-  border: solid 2px black;
-  border-radius: 10px;
-  margin: 2%;
-}
-
-.machine {
-  border: solid 2px black;
-  width: 60px;
-  height: 60px;
-  border-radius: 10px;
-  margin: 2%;
-}
-
-.machine:hover, .action:hover {
-  border: solid 2px orange;
-}
-
-
-.machinesBox {
-  border-radius: 5px;
-  height: 50%;
-  background-color: #E6E4E4;
-  margin: 10%;
-  border: solid 5px #E6E4E4;
-}
-
-.actionsBox {
-  border-radius: 5px;
-  height: 50%;
-  background-color: #E6E4E4;
-  margin: 10%;
-  border: solid 5px #E6E4E4;
-}
-
-.toolboxMenu {
-  background-color: #7D7979;
-  width: 15%;
-  align-content: center;
-  border: solid 5px #7D7979; 
-  float: left;
-}
-
-.toolboxTitle {
-  color: orange;
-  margin-left: 5px;
-}
-
-.grid {
-  border: solid 5px #7D7979;
-  float: left;
-}
-
-.empty {
-  height: 60px;
-  width: 60px;
-  border: solid 3px black;
-}
