+++ front/src/index.jsx
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 
-import './index.css';
 import App from './components/app/App';
 import reducers from './reducers/toolboxReducer';
 import * as serviceWorker from './serviceWorker';
